Issue #72 Update karma to INT8 (as signed int caused build errors)
Remove show_number_on_screen due to errors about length
-e 
Original/build commit ID: 1472516880ab09e40bc6ff3b7a2440e98a2e3d92

+++ game_state.c
@@ -76,7 +76,7 @@ typedef struct {
     unsigned int intelligence;
     unsigned int strength;
     unsigned int charm;
-    int karma;
+    INT8 karma;
 
     // Money the user has.
     unsigned int balance;
